Migracion Tabla sucursal, cambio atributo telefono_sucursal

+++ database/migrations/2020_10_31_214027_create_sucursal_table.php
@@ -20,7 +20,7 @@ class CreateSucursalTable extends Migration
             $table->string('descripcion_sucursal', 250);
             $table->string('cufe_sucursal')->unique();
             $table->string('email_sucursal');
-            $table->integer('telefono_sucursal');
+            $table->bigInteger('telefono_sucursal');
             $table->string('direccion_sucursal', 250);
             $table->boolean('habilitado');
             $table->boolean('borrado_logico_sucursal');
